Updated the auxiliary reactors base class
Auxiliary reactors should now automatically disable themselves if the
ship they are installed and activated on has less fuel in it than what
the reactor would consume in two onUpdate events.

The changelog is now separated from the general readme and each
individual update listed in the changelog now has a line of hyphens
underneath it to make it easier to spot in a quick glance.