Updated, changed, and fixed descriptions of items
Also updated to API v28 for Transcendence v1.6.3 (nothing had to
actually be changed for this).
Updated whitespace and commenting standard for stuff (this isn't
finished or documented yet).
Military crew members are now identified by an attribute that won't tend
to get them arrested.
See changelog for more details.